requirements.txt
awscrt==0.23.0
brotli==1.1.0
brotlicffi==1.1.0.0
ca_certs_locater==1.0
cffi==1.16.0
chardet==4.0.0
click==8.0.3
cryptography==3.4.8
docutils==0.21.2
google_api_python_client==2.149.0
h2==4.1.0
ipython==8.12.3
Jinja2==3.1.2
js==1.0
oauth2client==4.1.3
openai==1.54.4
packaging==24.2
pyodide==0.0.2
pyOpenSSL==24.2.1
pyu2f==0.1.5
railroad==0.5.0
simplejson==3.19.3
Sphinx==8.1.3
typing_extensions==4.12.2
youtube-transcript-api>=1.2.2
zstandard==0.23.0
